end_to_end_AI_Agentic_news-research-writing-system/requirements.txt
crewai
langchain-google-genai # Used to interact with Google gemini models
load_dotenv
crewai_tools

